Feat(plot) pie chart chart for occhab
